Remove hack that now breaks on Python 3.4+
The hack was introduced in b4a199597fd65052f72d9d49f0e43c25004a8a74
because zope.testrunner didn't have a __main__.py, but now it has one,
so it's not necessary.

I don't know why the hack broke with the release of coverage 4.0, and
only on Python 3.4 and 3.5, and I don't want to know.