Unbreak 'make test'
Only spawn pager when stderr is a tty (would love to check stdout, but
stdout is *not* a tty inside $(shell ...) sections!), and $TERM is not
dumb.

Don't force ANSI color codes when $TERM is dumb or stderr is not a
terminal.

Make sure to exit with a non-zero when 'make test' fails and you pipe
the errors through a pager.